Add interval/sphgeom-style APIs to Boxes.
These mostly expose functionality the Boxes already expose some other
way, but consistent interfaces are important.

Despite the fact that Boxes are already mutable in Python, I have not
exposed the interval/sphgeom-style mutation APIs, in case we decide
we want to make Box immutable in Python in the future.